+++ Dungeon/src/main/java/dungeon/domain/Bite.java
@@ -6,7 +6,7 @@ package dungeon.domain;
 public class Bite implements Attack {
 
     @Override
-    public int apply(Actor source, Actor target) {
+    public double apply(Actor source, Actor target) {
         return target.damage(source.getMaxHealth() / 10 * 2);
     }
