+++ Dungeon/src/main/java/dungeon/domain/Punch.java
@@ -7,7 +7,7 @@ package dungeon.domain;
 public class Punch implements Attack {
 
     @Override
-    public int apply(Actor source, Actor target) {
+    public double apply(Actor source, Actor target) {
         return target.damage(source.getMaxHealth() / 10 * 4);
     }
